Increase global test coverage to >99% (#376)
* Add vault tests

* Test updateCentrifugeGasPrice

* Change last computed at to uint64

* Fix testUpdateCentrifugeGasPrice

* Disable fork tests

* Add erc20 file test

* Add invalid message handle tests

* Update invalid message tests

* Redeploy contracts

* Add zero price check

* Redeploy router

* Add deposit test

* Add reentrancy tests

* Add multicall reentrancy mock

* Cleanup

* Update many docs

* Update src/interfaces/token/ITranche.sol


---------

+++ src/interfaces/gateway/IGasService.sol
@@ -28,7 +28,7 @@ interface IGasService {
     /// Keeps track what the last time was when the gas price was updated
     /// @dev This is a getter method
     /// @return Timestamp when the gas price was last updated
-    function lastUpdatedAt() external returns (uint256);
+    function lastUpdatedAt() external returns (uint64);
 
     /// CFG/ETH price
     /// @dev This is a getter method
@@ -37,7 +37,7 @@ interface IGasService {
 
     /// Called to update the  gas price
     /// @param value New price in Centrifuge Chain base unit
-    function updateGasPrice(uint128 value, uint256 computedAt) external;
+    function updateGasPrice(uint128 value, uint64 computedAt) external;
 
     /// Called to update the  CFG/ETH price
     /// @param value New price in wei
